divide launchpad
