+++ src/front/interface/launchpad.js
@@ -4,7 +4,6 @@
 
 const $ = require('jquery');
 const Editor = require('./aceditor');
-const Stdios = require('./stdios');
 
 
 // メモ化
@@ -16,33 +15,11 @@ function m$(html) {
         (_m_memorized[html] = $(html));
 }
 
-// _____________________________________________________
-// initialize
-
-$(() => {
-    $('textarea.enabletabs').keydown((e) => {
-        if (e.keyCode === 9) {
-            e.preventDefault(); // デフォルト動作の中止
-            var elem = e.target;
-            var val = elem.value;
-            var pos = elem.selectionStart;
-            elem.value = val.substr(0, pos) + '\t' + val.substr(pos, val.length);
-            elem.setSelectionRange(pos + 1, pos + 1);
-        }
-    });
-});
-
 
 // _____________________________________________________
 // getter
 
-function getSelectedTask() { // TODO: moved TaskSelector
-    return $('#selector_codelang option:selected').data('cmd');
-}
-
-
-export function gatherInfo() {
-    const cmd = getSelectedTask();
+export function getOptionValues(cmd) { // TODO: moved TaskSelector
     const options = {};
     $('#div_options > div')
         .filter((i, e) => ($(e).data('cmd') == cmd))
@@ -50,14 +27,9 @@ export function gatherInfo() {
         .each((i, e) => {
             options[$(e).data('key')] = $(e).val();
         });
-
-    return {
-        txt_stdins: Stdios.getStdins(true),
-        txt_code: Editor.getValue(),
-        cmd: cmd,
-        options: options
-    };
+    return options;
 }
+
 
 
 // _____________________________________________________
@@ -72,15 +44,6 @@ export function changeVisibleRecipes(cmd, lang) {
 }
 
 
-export function displayStdout(text, id) { // TODO: ????
-    if (id) {
-        const li = {};
-        li[id] = text;
-        Stdios.setStdouts(li);
-    }
-}
-
-
 // _____________________________________________________
 // setup
 
@@ -89,6 +52,11 @@ export function addClickRecipeListener(handler) {
     clickRecipeHandlers.push(handler);
 }
 
+
+export function appendTask(taskInfo) {
+    appendRecipes(taskInfo);
+    appendTaskOptions(taskInfo);
+}
 
 
 function appendRecipes(taskInfo) {
@@ -132,10 +100,4 @@ function appendTaskOptions(taskInfo) {
         );
     }
     m$('#div_options').append(domc);
-}
-
-
-export function appendTask(taskInfo) { // TODO: appendTask
-    appendRecipes(taskInfo);
-    appendTaskOptions(taskInfo);
 }
